@@ -13,7 +13,7 @@ class ZipcodeController extends Controller
     //
     public function search_zipcode($code)
     {
-        
+
         /* search exact city zip code information */
         $sql  = DB::table('postal_codes')
             ->where('postal', '=', $code)
@@ -25,11 +25,11 @@ class ZipcodeController extends Controller
                 'phone_area_code'
             )
             ->get();
-        
+
         /*  */
         $page_title    = sprintf('%s Zip Code', $code);
         $page_info     = sprintf('%s Zip Code is not yet in our database.', $code);
-        
+
         if ( count($sql) > 0 ) {
 
             /* Zip Code is belong to more than 1 barangay */
@@ -51,53 +51,54 @@ class ZipcodeController extends Controller
 
                                         'barangay'      => $value->barangay,
                                         'barangay_url'  => route( 'url_barangay',   [
-                                                                                        'city' => str_replace(' ', '-', strtolower($value->city)), 
+                                                                                        'city' => str_replace(' ', '-', strtolower($value->city)),
                                                                                         'barangay' => str_replace(' ', '-', strtolower($value->barangay))
                                                                                     ]),
-                                                    
+
                                         'postal'        => $value->postal,
                                         'postal_url'    => route( 'url_zipcode',    [
-                                                                                        'code' => str_replace(' ', '-', strtolower($value->postal)), 
+                                                                                        'code' => str_replace(' ', '-', strtolower($value->postal)),
                                                                                     ]),
 
                                         'city'          => $value->city,
                                         'city_url'      => route( 'url_city',   [
                                                                                     'city' => str_replace(' ', '-', strtolower($value->city))
                                                                                 ]),
-                                        
+
                                         'phone_area_code' => $value->phone_area_code
                                     ];
 
             }
-    
+
             /*  */
             $data    = array(
-                        'page_title'    => $page_title,
-                        'canonical'     => route('url_zipcode', ['code' => $code]),
-                        'description'   => $page_info,
-
-                        'page_info'     => $page_info,
-                        'results'       => json_decode(json_encode($query_data)), /* <- convert array into object */
-                        'search_q'      => '',
+                        'page_title'        => $page_title,
+                        'canonical'         => route('url_zipcode', ['code' => $code]),
+                        'description'       => $page_info,
+                        'subheader_title'   => $page_title,
+                        'page_info'         => $page_info,
+                        'results'           => json_decode(json_encode($query_data)), /* <- convert array into object */
+                        'search_q'          => '',
                     );
 
         }
-        else { 
+        else {
 
             /*  */
             $data    = array(
-                        'page_title'    => $page_title,
-                        'canonical'     => null,
-                        'description'   => $page_info,
+                        'page_title'        => $page_title,
+                        'canonical'         => null,
+                        'description'       => $page_info,
+                        'subheader_title'   => $page_title,
 
-                        'page_info'     => $page_info,
-                        'results'       => null,
-                        'search_q'      => '',
+                        'page_info'         => $page_info,
+                        'results'           => null,
+                        'search_q'          => '',
                     );
 
         }
 
         return view('pages.zipcodes.zipcode', $data);
-        
+
     }
 }
