@@ -1,94 +1,48 @@
 @extends('layouts.app')
 @section('title', $page_title)
-
-
-
 @section('page_styles')
-<link rel="stylesheet" type="text/css" href="https://cdn.datatables.net/v/bs4/dt-1.10.22/fc-3.3.1/fh-3.1.7/r-2.2.6/datatables.min.css"/>
+
 @endsection
-
-
 
 @section('content')
-<div class="container mt-3">
+<div class="relative overflow-hidden py-10 md:py-5">
 
-    <!-- ads -->
-    @include('ads.ads1')
-    <!--  -->
-    <!--  -->
+    <div class="mx-auto max-w-7xl sm:px-6 lg:px-8">
 
-
-
-    <!-- search -->
-    @include('includes.search')
-    <!--  -->
-    <!--  -->
-
-
-
-    @if ( null !== $results )
-
-        <div class="row mb-2">
-            <div class="col-xl-12">
-                <h1 class="font-weight-light f-16">{{ $page_info }}</h1>
-            </div>
-        </div>  
-
-        <div class="table-responsive-md">
-            <table class="table" id="results-table">
-                <thead>
-                    <tr style="white-space:nowrap;">
-                        <th scope="col">Zip Code</th>
-                        <th scope="col">Barangay</th>
-                        <th scope="col">City</th>
-                        <th scope="col">Region</th>
-                        <th scope="col">IDD: area code</th>
-                    </tr>
-                </thead>
-                <tbody>
-                    @foreach($results as $idx => $value)
-                    <tr>
-                        <td>
-                            <a href="{{ $value->postal_url }}">{{ $value->postal  }}</a>
-                        </td>
-                        <td>
-                            <a href="{{ $value->barangay_url }}">{{ $value->barangay  }}</a>
-                        </td>
-                        <td>
-                            <a href="{{ $value->city_url }}">{{ $value->city  }}</a>
-                        </td>
-                        <td>
-                            <a href="{{ urldecode($value->url_region) }}">{{ $value->region  }}</a>
-                        </td>
-                        <td>
-                            {{ $value->phone_area_code  }}
-                        </td>
-                    </tr>
-                    @endforeach
-                </tbody>
-            </table>
-        </div> <!-- [ .table-responsive ] END -->
-    
-    @else
-
-        <!-- no results -->
-        @include('includes.no-results')
+        <!-- search -->
+        @include('includes.search')
         <!--  -->
         <!--  -->
 
-    @endif
+        <!-- ads -->
+        @include('ads.ads1')
+        <!--  -->
+        <!--  -->
 
-    <!-- ads -->
-    @include('ads.ads2')
-    <!--  -->
-    <!--  -->
+        @if ( null !== $results )
+            {{-- zip code list display --}}
+            <x-zip-list :results="$results" :page_info="$page_info" />
+        @else
 
-</div> <!-- [ .container ] END -->
+            <!-- no results -->
+            @include('includes.no-results')
+            <!--  -->
+            <!--  -->
+
+        @endif
+
+        <!-- ads -->
+        @include('ads.ads2')
+        <!--  -->
+        <!--  -->
+
+    </div>
+</div>
 @endsection
 
 
 
-@section('page_scripts') 
+@section('page_scripts')
 <!-- dataTables -->
 <script type="text/javascript" src="https://cdn.datatables.net/v/bs4/dt-1.10.22/fc-3.3.1/fh-3.1.7/r-2.2.6/datatables.min.js"></script>
 <script type="text/javascript" src="https://cdn.jsdelivr.net/npm/datatables.mark.js@2.0.2/dist/datatables.mark.es6.min.js"></script>
